+++ Sources/TuistGenerator/Generator/LinkGenerator.swift
@@ -241,7 +241,7 @@ final class LinkGenerator: LinkGenerating {
             .map { LinkGeneratorPath.absolutePath($0.removingLastComponent()) }
         let sdkPaths = dependencies.compactMap { (dependency: GraphDependencyReference) -> LinkGeneratorPath? in
             if case let GraphDependencyReference.sdk(_, _, source) = dependency {
-                return LinkGeneratorPath.string(source.frameworkSearchPath)
+                return source.frameworkSearchPath.map { LinkGeneratorPath.string($0) }
             } else {
                 return nil
             }
@@ -338,7 +338,7 @@ final class LinkGenerator: LinkGenerating {
                 switch dependency {
                 case let .framework(path, _, _, _, _, _, _, _):
                     try addBuildFile(path)
-                case let .library(path, _, _, _, _):
+                case let .library(path, _, _, _):
                     try addBuildFile(path)
                 case let .xcframework(path, _, _, _):
                     try addBuildFile(path)